+++ vue-components/src/main.ts
@@ -10,6 +10,8 @@ import Dropdown from './components/Dropdown.vue';
 import Message from './components/Message.vue';
 import Popover from './components/Popover.vue';
 import Icon from './components/Icon.vue';
+import BouncingDots from './components/BouncingDots.vue';
+import QuantityInput from './components/QuantityInput.vue';
 
 export {
 	Button,
@@ -24,4 +26,6 @@ export {
 	Message,
 	Popover,
 	Icon,
+	BouncingDots,
+	QuantityInput,
 };